rfcomm: Fix a slot initform
rfcomm-socket did not correctly override the family slot in the socket
superclass, causing an erroneous "No socket family" error to occur
during a make-instance call for the rfcomm-socket.

The fix was to prefix the slot name with the package name to force CLOS
to override that slot, rather than a slot name specific to the sb-bluez
package. Weird.